Tests for Pistache errors
