+++ src/util/files.cpp
@@ -1,12 +1,6 @@
 #include "files.h"
 #include "bytes.h"
 #include "logging.h"
-#include "pistache/async.h"
-#include "pistache/http_header.h"
-
-#include <pistache/client.h>
-#include <pistache/http.h>
-#include <pistache/net.h>
 
 #include <cstdlib>
 #include <cstring>
@@ -14,8 +8,6 @@
 #include <iostream>
 #include <sstream>
 #include <string>
-
-#define HTTP_FILE_TIMEOUT 20000
 
 namespace faabric::util {
 std::string readFileToString(const std::string& path)
@@ -71,53 +63,6 @@ size_t writeDataCallback(void* ptr, size_t size, size_t nmemb, void* stream)
     std::string data((const char*)ptr, (size_t)size * nmemb);
     *((std::stringstream*)stream) << data;
     return size * nmemb;
-}
-
-std::vector<uint8_t> readFileFromUrl(const std::string& url)
-{
-    return readFileFromUrlWithHeader(url, nullptr);
-}
-
-std::vector<uint8_t> readFileFromUrlWithHeader(
-  const std::string& url,
-  const std::shared_ptr<Http::Header::Header>& header)
-{
-    auto logger = faabric::util::getLogger();
-
-    Http::Client client;
-    client.init();
-
-    auto rb = client.get(url);
-
-    if (header != nullptr) {
-        rb.header(header);
-    }
-
-    Async::Promise<Http::Response> resp = rb.send();
-
-    std::stringstream out;
-    resp.then(
-      [&](Http::Response response) {
-          if (response.code() == Http::Code::Ok) {
-              auto body = response.body();
-              out << body;
-          } else {
-              logger->error(
-                "Failed request to {}, code = {}", url, response.code());
-          }
-      },
-      [&](std::exception_ptr exc) {
-          PrintException excPrinter;
-          excPrinter(exc);
-      });
-
-    Async::Barrier<Http::Response> barrier(resp);
-    std::chrono::milliseconds timeout(HTTP_FILE_TIMEOUT);
-    barrier.wait_for(timeout);
-
-    client.shutdown();
-
-    return faabric::util::stringToBytes(out.str());
 }
 
 bool isWasm(const std::vector<uint8_t>& bytes)
